Back/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.108.0"
uvicorn = "0.25.0"
httpie = "3.2.2"
requests = "2.31.0"
httpx = "0.26.0"
pydantic = "2.5.3"
sqlmodel = "^0.0.14"
pydantic-settings = "^2.1.0"
psycopg2 = "^2.9.9"
gunicorn = "^22.0.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
python-multipart = "^0.0.20"
alembic = "^1.14.1"
email-validator = "^2.2.0"
bcrypt = "4.0.1"
passlib = "1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
